requirements-scripts.txt
# Requirements for Python scripts
PyPDF2==3.0.1
supabase==2.0.0
python-dotenv==1.0.0